requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
Faker==19.6.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
